// ng-jz-paint/src/app/interfaces/settings-interface.ts
export interface SettingsInterface {

    wallPricePerSqft: string;
    ceilingPricePerSqft: string;
    trimPricePerSqft: string;
    pricePerDoorFace: string;
    pricePerDoorFrame: string;
    pricePerWindowFrame: string;
    timeToPaymentDue: string;
    percentageLateFee: string;

    paymentStatusID: string;

}